This is a very big revision with several important and subtle changes.
The script to test learned constraints has been improved.

One change is that we detect changes in variable state differently, now using the
depth of the trail stack combined with the state of the variable bounds.

Added a special check for failure at the root node.

Ensure that unary constraints store explanations correctly.

Make sure explanations for internal events in long int var, such as assignments
due to MHAV, are stored correctly. There was a bug before that meant they were
filtered out as being duplicate explanations when they were not.

The remaining changes are to accomodate a new type of failure (AMOV), where we experience
a double assignment to the same variable. Storing an assignment explanation should
no longer be predicated on it being currently unassigned, since we now might store a second.
Next, in the variable types there is now space for two assignment explanations,
and the getters and setters have been amended accordingly. The addition of a AMOV failure
is not strictly necessary, since the MHAV nogood could take over, what is more important
is the ability to store two assignment explanations.

Finally I have removed the test in distribute() to avoid zero things being distributed,
I can't remember why.


git-svn-id: svn+ssh://keith.cs.st-and.ac.uk/mnt/raid/svn/repositories/minion-private/branches/learning@2288 477b24dd-e91e-0410-aa49-980f3df0729b