Revert "attempt at throwing away old restart nogoods when splitting multiple times"
This reverts commit abb15f0eb4fd7a97f4092c54bed1ff3d816b0ea1.

Throwing away the old restart nogoods does not work because the new ones do not
necessarily subsume them. The new nogoods will get minion to where it left off
/from the node where it was stopped the previous time/. The old nogoods are not
encoded in the new ones because getting to the node is now not a search decision
(which would be encoded) but achieved through propagation of the old noogoods.